refactor: extract shop logic into composables
